refactor(preferences): 拆分为 preferences/ 子模块（model/normalize/categoryPrefs/customSources/settings）
- model.ts: 偏好类型、默认值、排版/预存选项表与共享校验
- normalize.ts: normalizePreferences 持久化数据归一化
- categoryPrefs.ts: 分类顺序/显隐/信源覆盖查询与不可变更新、自建分类 CRUD
- customSources.ts: 自定义订阅源增删改与 OPML 批量导入
- settings.ts: 主题/配色/墨水屏/预存/排版/代理 setter
- preferences.ts 保留原路径作为 barrel，公开 API 不变
